Add preview to request view
